create new functional file
